add: proxy test factory
