+++ lib/github.js
@@ -13,7 +13,7 @@ module.exports = function github ({ url, auth }) {
   }
 
   async function getPRLabels (prId) {
-    const response = await octokit.issues.listLabelsOnIssue({ owner, repo: name, number: prId })
+    const response = await octokit.issues.listLabelsOnIssue({ owner, repo: name, issue_number: prId })
     response.pullRequest = { id: prId }
     return response
   }
