+++ lib/npm.js
@@ -7,6 +7,7 @@ function runSpawn (cwd, cmd, args) {
     const cli = spawn(cmd, args, { cwd, env: process.env, shell: true })
     cli.stdout.setEncoding('utf8')
 
+    // TODO these streams are useful for debug!
     // cli.stdout.pipe(process.stdout)
     // cli.stderr.pipe(process.stdout)
